replace manual network layer to autogenerated

+++ app/src/main/kotlin/com/example/popularmovies/domain/MoviesRepository.kt
@@ -8,6 +8,6 @@ import kotlinx.coroutines.flow.Flow
 interface MoviesRepository {
   val popularMovies: Flow<List<MoviePreview>>
   fun observeMovieDetailsById(id: MovieId): Flow<MovieDetails?>
-  suspend fun fetchPopularMovies(language: String, page: Int, region: String)
+  suspend fun fetchPopularMovies(language: String, page: Long, region: String)
   suspend fun fetchMovieDetailsById(id: MovieId, language: String)
 }
